+++ src/CSE_CST328_Driver.h
@@ -9,7 +9,7 @@
   Version: 0.0.1
   License: MIT
   Source: https://github.com/CIRCUITSTATE/CSE_Touch
-  Last Modified: +05:30 12:36:47 PM 09-03-2025, Sunday
+  Last Modified: +05:30 15:33:34 PM 09-03-2025, Sunday
  */
 //======================================================================================//