+++ Changes.md
@@ -1,6 +1,15 @@
 
 
 #
+### **+05:30 07:58:07 PM 10-03-2025, Monday**
+
+  - Added PlatformIO examples to the `extras` folder.
+    - Added `Read-Touch-Polling` example.
+    - Added `Read-Touch-Interrupt` example.
+    - Both examples tested with FireBeetle-ESP32E board.
+  - Updated Arduino library specifications.
+
+#
 ### **+05:30 12:37:13 PM 09-03-2025, Sunday**
 
   - Updated Readme.